Make reusable a component of AddCardScreen.kt #13
